+++ src/main/scala/com/microsoft/hyperspace/util/JsonUtils.scala
@@ -16,16 +16,15 @@
 
 package com.microsoft.hyperspace.util
 
-import scala.math.Ordered.orderingToOrdered
-
 import com.fasterxml.jackson.annotation.JsonInclude.Include
-import com.fasterxml.jackson.databind.{JsonDeserializer, JsonSerializer, ObjectMapper}
+import com.fasterxml.jackson.core.{JsonGenerator, JsonParser}
+import com.fasterxml.jackson.databind.{DeserializationContext, ObjectMapper, SerializerProvider}
+import com.fasterxml.jackson.databind.deser.std.StdDeserializer
 import com.fasterxml.jackson.databind.module.SimpleModule
+import com.fasterxml.jackson.databind.ser.std.StdSerializer
 import com.fasterxml.jackson.hyperspace.module.scala.ScalaObjectMapper
 import com.fasterxml.jackson.module.scala.DefaultScalaModule
-import org.apache.spark.sql.types.StructType
-
-import com.microsoft.hyperspace.shim.{StructTypeDeserializer, StructTypeSerializer}
+import org.apache.spark.sql.types.{DataType, StructType}
 
 /**
  * Useful json functions used around the Hyperspace codebase.
@@ -36,6 +35,8 @@ object JsonUtils {
   private val objectMapper = new ObjectMapper() with ScalaObjectMapper
   objectMapper.setSerializationInclusion(Include.ALWAYS)
   objectMapper.registerModule(DefaultScalaModule)
+
+  // Only needed for Spark 2.4
   objectMapper.registerModule {
     new SimpleModule()
       .addSerializer(classOf[StructType], new StructTypeSerializer)
@@ -52,5 +53,20 @@ object JsonUtils {
 
   def jsonToMap(json: String): Map[String, Any] = {
     objectMapper.readValue[Map[String, Any]](json)
+  }
+
+  private class StructTypeSerializer extends StdSerializer[StructType](classOf[StructType]) {
+    override def serialize(
+        value: StructType,
+        gen: JsonGenerator,
+        provider: SerializerProvider): Unit = {
+      gen.writeRawValue(value.json)
+    }
+  }
+
+  private class StructTypeDeserializer extends StdDeserializer[StructType](classOf[StructType]) {
+    override def deserialize(p: JsonParser, ctx: DeserializationContext): StructType = {
+      DataType.fromJson(p.readValueAsTree().toString).asInstanceOf[StructType]
+    }
   }
 }
